refactor: replace walletconnect.events with eventemitter.net
