--- Core Modules/WalletConnectSharp.Events/Interfaces/IEvents.cs
@@ -1,16 +1,0 @@
-using System;
-using WalletConnectSharp.Events.Model;
-
-namespace WalletConnectSharp.Events.Interfaces
-{
-    /// <summary>
-    /// An interface that represents a class that triggers events that can be listened to.
-    /// </summary>
-    public interface IEvents
-    {
-        /// <summary>
-        /// The EventDelegator that should be used to listen to (or trigger) events
-        /// </summary>
-        EventDelegator Events { get; }
-    }
-}
