--- Core Modules/WalletConnectSharp.Events/WalletConnectSharp.Events.csproj
@@ -1,30 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFrameworks>$(DefaultTargetFrameworks)</TargetFrameworks>
-        <PackageVersion>$(DefaultVersion)</PackageVersion>
-        <AssemblyVersion>$(DefaultVersion)</AssemblyVersion>
-        <FileVersion>$(DefaultVersion)</FileVersion>
-        <PackageId>WalletConnect.Events</PackageId>
-        <Title>WalletConnectSharp.Events</Title>
-        <Authors>pedrouid, gigajuwels, edkek</Authors>
-        <Description>A port of the TypeScript SDK to C#. A complete implementation of the WalletConnect v2 protocol that can be used to connect to external wallets or connect a wallet to an external Dapp</Description>
-        <Copyright>Copyright (c) WalletConnect 2023</Copyright>
-        <PackageProjectUrl>https://walletconnect.org/</PackageProjectUrl>
-        <PackageIcon>icon.png</PackageIcon>
-        <RepositoryUrl>https://github.com/WalletConnect/WalletConnectSharp</RepositoryUrl>
-        <RepositoryType>git</RepositoryType>
-        <PackageTags>walletconnect wallet web3 ether ethereum blockchain evm</PackageTags>
-        <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-        <PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\WalletConnectSharp.Common\WalletConnectSharp.Common.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <None Include="..\..\resources\icon.png" Pack="true" PackagePath="\" />
-    </ItemGroup>
-
-</Project>
