+++ Core Modules/WalletConnectSharp.Network/Interfaces/IJsonRpcConnection.cs
@@ -1,6 +1,4 @@
-using System;
-using System.Threading.Tasks;
-using WalletConnectSharp.Events.Interfaces;
+using EventEmitter.NET;
 
 namespace WalletConnectSharp.Network
 {
