+++ Core Modules/WalletConnectSharp.Network/JsonRpcProvider.cs
@@ -1,9 +1,10 @@
+using EventEmitter.NET;
+using EventEmitter.NET.Model;
 using Newtonsoft.Json;
 using WalletConnectSharp.Common;
 using WalletConnectSharp.Common.Logging;
 using WalletConnectSharp.Common.Model.Errors;
-using WalletConnectSharp.Events;
-using WalletConnectSharp.Events.Model;
+using WalletConnectSharp.Common.Utils;
 using WalletConnectSharp.Network.Models;
 
 namespace WalletConnectSharp.Network
@@ -94,7 +95,7 @@ namespace WalletConnectSharp.Network
         public JsonRpcProvider(IJsonRpcConnection connection)
         {
             _context = Guid.NewGuid();
-            this._delegator = new EventDelegator(this);
+            this._delegator = new EventDelegator(Context);
             this._connection = connection;
             if (this._connection.Connected)
             {
