+++ Tests/WalletConnectSharp.Auth.Tests/AuthClientTest.cs
@@ -7,10 +7,7 @@ using WalletConnectSharp.Common.Utils;
 using WalletConnectSharp.Core;
 using WalletConnectSharp.Core.Models.Pairing;
 using WalletConnectSharp.Core.Models.Publisher;
-using WalletConnectSharp.Core.Models.Relay;
 using WalletConnectSharp.Core.Models.Verify;
-using WalletConnectSharp.Events;
-using WalletConnectSharp.Network.Models;
 using WalletConnectSharp.Storage;
 using WalletConnectSharp.Tests.Common;
 using Xunit;
