+++ Tests/WalletConnectSharp.Network.Tests/WalletConnectSharp.Network.Tests.csproj
@@ -22,7 +22,6 @@
 
     <ItemGroup>
       <ProjectReference Include="..\..\Core Modules\WalletConnectSharp.Common\WalletConnectSharp.Common.csproj" />
-      <ProjectReference Include="..\..\Core Modules\WalletConnectSharp.Events\WalletConnectSharp.Events.csproj" />
       <ProjectReference Include="..\..\Core Modules\WalletConnectSharp.Network.Websocket\WalletConnectSharp.Network.Websocket.csproj" />
       <ProjectReference Include="..\..\Core Modules\WalletConnectSharp.Network\WalletConnectSharp.Network.csproj" />
       <ProjectReference Include="..\WalletConnectSharp.Tests.Common\WalletConnectSharp.Tests.Common.csproj" />
