+++ WalletConnectSharp.Core/Controllers/HeartBeat.cs
@@ -1,6 +1,7 @@
+using EventEmitter.NET;
+using WalletConnectSharp.Common.Utils;
 using WalletConnectSharp.Core.Interfaces;
 using WalletConnectSharp.Core.Models.Heartbeat;
-using WalletConnectSharp.Events;
 
 namespace WalletConnectSharp.Core.Controllers
 {
@@ -60,7 +61,7 @@ namespace WalletConnectSharp.Core.Controllers
         /// <param name="interval">The interval to emit the <see cref="IHeartBeat.OnPulse"/> event at</param>
         public HeartBeat(int interval = 5000)
         {
-            Events = new EventDelegator(this);
+            Events = new EventDelegator(Context);
 
             Interval = interval;
         }
