+++ WalletConnectSharp.Core/Controllers/Publisher.cs
@@ -1,10 +1,9 @@
+using EventEmitter.NET;
 using WalletConnectSharp.Common.Model.Relay;
 using WalletConnectSharp.Common.Utils;
 using WalletConnectSharp.Core.Interfaces;
-using WalletConnectSharp.Core.Models.Heartbeat;
 using WalletConnectSharp.Core.Models.Publisher;
 using WalletConnectSharp.Core.Models.Relay;
-using WalletConnectSharp.Events;
 using WalletConnectSharp.Network.Models;
 
 namespace WalletConnectSharp.Core.Controllers
@@ -58,7 +57,7 @@ namespace WalletConnectSharp.Core.Controllers
         public Publisher(IRelayer relayer)
         {
             Relayer = relayer;
-            Events = new EventDelegator(this);
+            Events = new EventDelegator(Context);
             
             RegisterEventListeners();
         }
