+++ WalletConnectSharp.Core/Controllers/Relayer.cs
@@ -1,3 +1,4 @@
+using EventEmitter.NET;
 using Newtonsoft.Json;
 using WalletConnectSharp.Common;
 using WalletConnectSharp.Common.Logging;
@@ -6,8 +7,6 @@ using WalletConnectSharp.Common.Utils;
 using WalletConnectSharp.Core.Interfaces;
 using WalletConnectSharp.Core.Models.Relay;
 using WalletConnectSharp.Core.Models.Subscriber;
-using WalletConnectSharp.Events;
-using WalletConnectSharp.Events.Model;
 using WalletConnectSharp.Network;
 using WalletConnectSharp.Network.Models;
 
@@ -128,7 +127,7 @@ namespace WalletConnectSharp.Core.Controllers
         public Relayer(RelayerOptions opts)
         {
             Core = opts.Core;
-            Events = new EventDelegator(this);
+            Events = new EventDelegator(Context);
             Messages = new MessageTracker(Core);
             Subscriber = new Subscriber(this);
             Publisher = new Publisher(this);
