+++ WalletConnectSharp.Core/Interfaces/IJsonRpcHistory.cs
@@ -1,10 +1,7 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
+using EventEmitter.NET;
 using WalletConnectSharp.Common;
 using WalletConnectSharp.Core.Models.History;
-using WalletConnectSharp.Events.Interfaces;
 using WalletConnectSharp.Network;
-using WalletConnectSharp.Network.Models;
 
 namespace WalletConnectSharp.Core.Interfaces
 {
