+++ WalletConnectSharp.Core/Interfaces/ISubscriber.cs
@@ -1,9 +1,7 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
+using EventEmitter.NET;
 using WalletConnectSharp.Common;
 using WalletConnectSharp.Core.Models.Relay;
 using WalletConnectSharp.Core.Models.Subscriber;
-using WalletConnectSharp.Events.Interfaces;
 
 namespace WalletConnectSharp.Core.Interfaces
 {
