+++ WalletConnectSharp.Sign/Interfaces/IEngineAPI.cs
@@ -1,6 +1,6 @@
+using EventEmitter.NET;
 using WalletConnectSharp.Core.Models.Pairing;
 using WalletConnectSharp.Core.Models.Relay;
-using WalletConnectSharp.Events.Interfaces;
 using WalletConnectSharp.Network.Models;
 using WalletConnectSharp.Sign.Models;
 using WalletConnectSharp.Sign.Models.Engine;
