+++ WalletConnectSharpV2.sln
@@ -6,15 +6,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WalletConnectSharp.Storage"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WalletConnectSharp.Crypto", "Core Modules\WalletConnectSharp.Crypto\WalletConnectSharp.Crypto.csproj", "{CE7D22CE-523B-47D1-8093-BB9643C4F6FB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WalletConnectSharp.Events", "Core Modules\WalletConnectSharp.Events\WalletConnectSharp.Events.csproj", "{FD172F4E-A33D-47F9-AD46-AD346F5B767B}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WalletConnectSharp.Network.Websocket", "Core Modules\WalletConnectSharp.Network.Websocket\WalletConnectSharp.Network.Websocket.csproj", "{53A4B228-6AFA-4BAE-B3CE-4C4711A8878B}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WalletConnectSharp.Network.Tests", "Tests\WalletConnectSharp.Network.Tests\WalletConnectSharp.Network.Tests.csproj", "{F237BAA4-EE57-4E34-91CA-37380FBA4D2F}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WalletConnectSharp.Common", "Core Modules\WalletConnectSharp.Common\WalletConnectSharp.Common.csproj", "{90B59A93-273C-49A3-8A26-37E487780B1A}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WalletConnectSharp.Events.Tests", "Tests\WalletConnectSharp.Events.Tests\WalletConnectSharp.Events.Tests.csproj", "{AFAA1188-AF03-4A79-A3A3-E25D08ED5757}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WalletConnectSharp.Storage.Test", "Tests\WalletConnectSharp.Storage.Test\WalletConnectSharp.Storage.Test.csproj", "{5EDF7334-9F47-4E0E-9964-6EA7934F0292}"
 EndProject
@@ -62,10 +58,6 @@ Global
 		{46EC1C68-C0C1-4515-B841-F117651C866E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{46EC1C68-C0C1-4515-B841-F117651C866E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{46EC1C68-C0C1-4515-B841-F117651C866E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FD172F4E-A33D-47F9-AD46-AD346F5B767B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FD172F4E-A33D-47F9-AD46-AD346F5B767B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FD172F4E-A33D-47F9-AD46-AD346F5B767B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FD172F4E-A33D-47F9-AD46-AD346F5B767B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{53A4B228-6AFA-4BAE-B3CE-4C4711A8878B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{53A4B228-6AFA-4BAE-B3CE-4C4711A8878B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{53A4B228-6AFA-4BAE-B3CE-4C4711A8878B}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -78,10 +70,6 @@ Global
 		{90B59A93-273C-49A3-8A26-37E487780B1A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{90B59A93-273C-49A3-8A26-37E487780B1A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{90B59A93-273C-49A3-8A26-37E487780B1A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AFAA1188-AF03-4A79-A3A3-E25D08ED5757}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AFAA1188-AF03-4A79-A3A3-E25D08ED5757}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AFAA1188-AF03-4A79-A3A3-E25D08ED5757}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AFAA1188-AF03-4A79-A3A3-E25D08ED5757}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5EDF7334-9F47-4E0E-9964-6EA7934F0292}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5EDF7334-9F47-4E0E-9964-6EA7934F0292}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5EDF7334-9F47-4E0E-9964-6EA7934F0292}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -129,12 +117,10 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{23897558-4177-425D-B2FA-75AB0B6FEDAE} = {3A55D6C7-0EF8-4EEA-90A5-89F5AFB97BA0}
-		{AFAA1188-AF03-4A79-A3A3-E25D08ED5757} = {3A55D6C7-0EF8-4EEA-90A5-89F5AFB97BA0}
 		{F237BAA4-EE57-4E34-91CA-37380FBA4D2F} = {3A55D6C7-0EF8-4EEA-90A5-89F5AFB97BA0}
 		{5EDF7334-9F47-4E0E-9964-6EA7934F0292} = {3A55D6C7-0EF8-4EEA-90A5-89F5AFB97BA0}
 		{90B59A93-273C-49A3-8A26-37E487780B1A} = {CDB4264E-3F30-4D4D-90BC-15E537BB739B}
 		{CE7D22CE-523B-47D1-8093-BB9643C4F6FB} = {CDB4264E-3F30-4D4D-90BC-15E537BB739B}
-		{FD172F4E-A33D-47F9-AD46-AD346F5B767B} = {CDB4264E-3F30-4D4D-90BC-15E537BB739B}
 		{0797C687-3123-4502-A441-F48A31E6EF46} = {CDB4264E-3F30-4D4D-90BC-15E537BB739B}
 		{53A4B228-6AFA-4BAE-B3CE-4C4711A8878B} = {CDB4264E-3F30-4D4D-90BC-15E537BB739B}
 		{46EC1C68-C0C1-4515-B841-F117651C866E} = {CDB4264E-3F30-4D4D-90BC-15E537BB739B}
